+++ src/main/java/kitchenpos/eatinorder/application/OrderTableService.java
@@ -1,6 +1,8 @@
 package kitchenpos.eatinorder.application;
 
 import kitchenpos.eatinorder.tobe.domain.*;
+import kitchenpos.eatinorder.tobe.domain.ordertable.OrderTable;
+import kitchenpos.eatinorder.tobe.domain.ordertable.OrderTableRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
