+++ src/main/java/kitchenpos/eatinorder/tobe/domain/ordertable/OrderTableRepository.java
@@ -1,4 +1,4 @@
-package kitchenpos.eatinorder.tobe.domain;
+package kitchenpos.eatinorder.tobe.domain.ordertable;
 
 import java.util.List;
 import java.util.Optional;
