+++ src/test/java/kitchenpos/eatinorder/tobe/domain/InMemoryOrderRepository.java
@@ -1,5 +1,7 @@
 package kitchenpos.eatinorder.tobe.domain;
 
+import kitchenpos.eatinorder.tobe.domain.ordertable.OrderTable;
+
 import java.util.*;
 
 public class InMemoryOrderRepository implements OrderRepository {
